python/viewer/requirements.txt
influxdb
pyyaml
tornado
